[FEAT] 뷰모델 반영 완
완!
