+++ lib/user/myPage/update_profile_page/update_profile_page.dart
@@ -1,12 +1,6 @@
-import 'dart:convert';
-import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:http/http.dart' as http;
-import 'package:http_parser/http_parser.dart';
-
-import '../../../service/location_service.dart';
+import 'package:mogu_app/user/myPage/update_profile_page/update_profile_page_viewModel.dart';
+import 'package:provider/provider.dart';
 
 class UpdateProfilePage extends StatefulWidget {
   final String userId;
@@ -31,118 +25,21 @@ class UpdateProfilePage extends StatefulWidget {
 }
 
 class _UpdateProfilePageState extends State<UpdateProfilePage> {
-  late String _nickname;
-  late String _profileImageUrl;
-  String? _address;
-  late double _latitude;
-  late double _longitude;
-  File? _newProfileImage;
-
   @override
   void initState() {
     super.initState();
-    _nickname = widget.nickname;
-    _profileImageUrl = widget.profileImage;
-    _latitude = widget.latitude;
-    _longitude = widget.longitude;
-    _loadAddress();
-  }
-
-  Future<void> _loadAddress() async {
-    final address = await LocationService().getAddressFromCoordinates(_latitude, _longitude);
-    setState(() {
-      _address = address;
-    });
-  }
-
-  Future<void> _openMapAndSelectLocation() async {
-    final selectedLocation = await LocationService().openMapPage(context);
-    if (selectedLocation != null) {
-      setState(() {
-        _latitude = selectedLocation.latitude;
-        _longitude = selectedLocation.longitude;
-      });
-      _loadAddress(); // 새 위치에 대한 주소를 로드
-    }
-  }
-
-  Future<void> _pickImage() async {
-    final ImagePicker _picker = ImagePicker();
-    final XFile? image = await _picker.pickImage(source: ImageSource.gallery);
-
-    if (image != null) {
-      setState(() {
-        _newProfileImage = File(image.path);
-      });
-    }
-  }
-
-  Future<void> _updateUser(BuildContext context) async {
-    if (_nickname == widget.nickname &&
-        _newProfileImage == null &&
-        _latitude == widget.latitude &&
-        _longitude == widget.longitude) {
-      Navigator.pop(context);
-      return;
-    }
-
-    String url = 'http://${dotenv.env['SERVER_IP']}:${dotenv.env['SERVER_PORT']}/user';
-    var request = http.MultipartRequest('PATCH', Uri.parse(url));
-
-    // Authorization 헤더에 토큰 추가
-    request.headers['Authorization'] = widget.token;
-
-    // JSON 데이터를 문자열로 변환하여 필드로 추가하고, Content-Type을 application/json으로 설정
-    request.fields['request'] = jsonEncode({
-      'nickname': _nickname,
-      'longitude': _longitude.toString(),
-      'latitude': _latitude.toString(),
-    });
-    request.headers['Content-Type'] = 'application/json';
-
-    // 이미지 파일이 있는 경우 파일 파트로 추가
-    if (_newProfileImage != null) {
-      request.files.add(
-        await http.MultipartFile.fromPath(
-          'image',
-          _newProfileImage!.path,
-          contentType: MediaType('image', 'jpeg'),
-        ),
-      );
-    }
-
-    try {
-      var streamedResponse = await request.send();
-      var response = await http.Response.fromStream(streamedResponse);
-
-      if (response.statusCode == 200) {
-        final responseData = jsonDecode(response.body);
-
-        // 서버로부터 새로운 프로필 이미지 URL을 받았다고 가정
-        final newProfileImageUrl = responseData['profileImageUrl'];
-        final newAddress = responseData['address'];
-
-        await _showSuccessDialog('성공', '프로필을 수정하였습니다.').then((_) {
-          Navigator.pop(context, {
-            'nickname': _nickname,
-            'profileImage': newProfileImageUrl,
-            'longitude': _longitude,
-            'latitude': _latitude,
-            'address': newAddress,
-          });
-        });
-      } else {
-        print('Failed with status code: ${response.statusCode}');
-        _showErrorDialog('오류', '서버에서 오류가 발생했습니다. 다시 시도해주세요.');
-      }
-    } catch (e) {
-      print('Exception: $e');
-      _showErrorDialog('오류', '서버와의 연결 오류가 발생했습니다. 다시 시도해주세요.');
-    }
+    final viewModel = Provider.of<UpdateProfilePageViewModel>(context, listen: false);
+    viewModel.initViewModel(
+      widget.nickname,
+      widget.profileImage,
+      widget.latitude,
+      widget.longitude
+    );
   }
 
   @override
   Widget build(BuildContext context) {
+    final viewModel = Provider.of<UpdateProfilePageViewModel>(context);
     return Scaffold(
       appBar: AppBar(
         leading: IconButton(
@@ -173,7 +70,7 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
       ),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
-        child: _address == null
+        child: viewModel.address == null
             ? Center(child: CircularProgressIndicator())
             : Column(
           crossAxisAlignment: CrossAxisAlignment.center,
@@ -191,12 +88,13 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
                       child: CircleAvatar(
                         radius: 40,
                         backgroundColor: Colors.grey.shade300,
-                        backgroundImage: _newProfileImage != null
-                            ? FileImage(_newProfileImage!)
-                            : (_profileImageUrl.isNotEmpty
-                            ? NetworkImage(_profileImageUrl) as ImageProvider
+                        backgroundImage: viewModel.newProfileImage != null
+                            ? FileImage(viewModel.newProfileImage!)
+                            : (viewModel.profileImageUrl.isNotEmpty
+                            ? NetworkImage(viewModel.profileImageUrl) as ImageProvider
                             : null),
-                        child: _profileImageUrl.isEmpty && _newProfileImage == null
+                        child: viewModel.profileImageUrl.isEmpty
+                            && viewModel.newProfileImage == null
                             ? Icon(Icons.camera_alt, size: 40, color: Colors.white)
                             : null,
                       ),
@@ -205,7 +103,7 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
                       top: 0,
                       right: 0,
                       child: GestureDetector(
-                        onTap: _pickImage,
+                        onTap: viewModel.pickImage,
                         child: Container(
                           padding: EdgeInsets.all(4.0),
                           decoration: BoxDecoration(
@@ -233,7 +131,7 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
                 Expanded(
                   child: InkWell(
                     onTap: () {
-                      _editNickname();
+                      viewModel.editNickname(context);
                     },
                     child: Container(
                       padding: EdgeInsets.symmetric(vertical: 6, horizontal: 10),
@@ -248,7 +146,7 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
                             child: Align(
                               alignment: Alignment.centerRight,
                               child: Text(
-                                _nickname,
+                                viewModel.nickname,
                                 style: TextStyle(
                                   fontSize: 18,
                                   fontWeight: FontWeight.w600,
@@ -271,21 +169,29 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
             ),
             SizedBox(height: 20),
             TextField(
-              controller: TextEditingController(text: _address),
+              controller: TextEditingController(text: viewModel.address),
               readOnly: true,
               decoration: InputDecoration(
                 labelText: '주소',
                 border: OutlineInputBorder(),
                 suffixIcon: IconButton(
                   icon: Icon(Icons.edit_location, color: Color(0xFFB34FD1)),
-                  onPressed: _openMapAndSelectLocation,
+                  onPressed: () {
+                    viewModel.openMapAndSelectLocation(context);
+                  },
                 ),
               ),
             ),
             SizedBox(height: 20),
             ElevatedButton(
               onPressed: () {
-                _updateUser(context);
+                viewModel.updateUser(
+                  context,
+                  widget.nickname,
+                  widget.latitude,
+                  widget.longitude,
+                  widget.token
+                );
               },
               style: ElevatedButton.styleFrom(
                 minimumSize: Size(double.infinity, 40),
@@ -300,79 +206,6 @@ class _UpdateProfilePageState extends State<UpdateProfilePage> {
           ],
         ),
       ),
-    );
-  }
-
-  void _editNickname() {
-    showDialog(
-      context: context,
-      builder: (context) {
-        TextEditingController controller = TextEditingController(text: _nickname);
-        return AlertDialog(
-          title: Text('닉네임 변경'),
-          content: TextField(
-            controller: controller,
-            decoration: InputDecoration(hintText: "새 닉네임을 입력하세요"),
-          ),
-          actions: [
-            TextButton(
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-              child: Text('취소'),
-            ),
-            TextButton(
-              onPressed: () {
-                setState(() {
-                  _nickname = controller.text;
-                });
-                Navigator.of(context).pop();
-              },
-              child: Text('확인'),
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future<void> _showSuccessDialog(String title, String message) {
-    return showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(title),
-          content: Text(message),
-          actions: <Widget>[
-            TextButton(
-              child: Text('확인'),
-              onPressed: () {
-                Navigator.of(context).pop(); // 다이얼로그를 닫습니다.
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-
-  Future<void> _showErrorDialog(String title, String message) {
-    return showDialog<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(title),
-          content: Text(message),
-          actions: <Widget>[
-            TextButton(
-              child: Text('확인'),
-              onPressed: () {
-                Navigator.of(context).pop(); // 다이얼로그를 닫습니다.
-              },
-            ),
-          ],
-        );
-      },
     );
   }
 }
